Add files via upload
This program is just a fun quiz to test an avengers fun on their movie knowledge. There is a bool array to store to make sure that the question either has or has not been asked yet, there are also vectors to store the questions that were wrong along with what the correct answer was, And what was the wrong answer. Has a preset of questions along with answer choices and the correct answer in their corresponding switch cases, although the computer will randomize which question goes next.